backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
gunicorn==21.2.0
python-multipart==0.0.6
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-core==1.29.5
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2